+++ app/Http/Controllers/Admin/ChatController.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace App\Http\Controllers\Admin;
+
+use App\Http\Controllers\GameController;
+use App\Models\Chat;
+use App\Models\User;
+
+class ChatController extends GameController
+{
+    public function index()
+    {
+        $this->restrictAccess(User::LEVEL_ADMIN);
+
+        return view('admin.chat', [
+            'messages' => Chat::get(),
+            'title' => trans('admin/chat.title'),
+        ]);
+    }
+
+    public function clear()
+    {
+        Chat::truncate();
+
+        return redirect()->route('admin.chat');
+    }
+
+    public function remove($id)
+    {
+        Chat::find($id)->delete();
+
+        return redirect()->route('admin.chat');
+    }
+}