+++ resources/lang/en/admin/chat.php
@@ -1,0 +1,12 @@
+<?php
+
+return [
+    'clear' => 'clear',
+    'count' => 'message(s) total',
+    'date' => 'Date',
+    'id' => 'ID',
+    'list-title' => 'List of messages',
+    'player' => 'Player',
+    'remove' => 'Remove',
+    'title' => 'Chat administration',
+];