+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Generally speaking, with the rewrite should not come any functional changes but
 where it was unavoidable or impractical to avoid the list item denotes most notable
 changes made when rewriting.
 
+- Admin: Chat administration
+    - Date format is now language dependent
 - Admin: Errors
     - Date format is now language dependent
 - Login